Add MachineExist to the plan spec

+++ internal/controller/plan_controller.go
@@ -37,6 +37,7 @@ import (
 	"easystack.com/plan/pkg/utils"
 	clusteropenstackapis "github.com/easystack/cluster-api-provider-openstack/api/v1alpha6"
 	clusteropenstackerrors "github.com/easystack/cluster-api-provider-openstack/pkg/utils/errors"
+	"github.com/go-logr/logr"
 	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack"
 	"github.com/gophercloud/gophercloud/openstack/compute/v2/extensions/servergroups"
@@ -123,7 +124,11 @@ type PlanMachineSetBind struct {
 //+kubebuilder:rbac:groups="",resources=secrets;,verbs=get;create;list;watch
 //+kubebuilder:rbac:groups="",resources=events,verbs=get;list;watch;create;update;patch
 
-// Reconcile is part of the main kubernetes reconciliation loop which aims to
+// Reconcile is part of the 
+
+
+
+kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // TODO(user): Modify the Reconcile function to compare the state specified by
 // the Plan object against the actual cluster state, and then
@@ -138,7 +143,6 @@ func (r *PlanReconciler) Reconcile(ctx context.Context, req ctrl.Request) (_ ctr
 		log      = log.FromContext(ctx)
 	)
 	// Fetch the OpenStackMachine instance.
-
 	plan := &ecnsv1.Plan{}
 	err := r.Client.Get(ctx, req.NamespacedName, plan)
 	if err != nil {
@@ -184,10 +188,6 @@ func (r *PlanReconciler) Reconcile(ctx context.Context, req ctrl.Request) (_ ctr
 			}
 		}
 	}
-	patchHelper, err := patch.NewHelper(plan, r.Client)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
 
 	osProviderClient, clientOpts, projectID, userID, err := provider.NewClientFromPlan(ctx, r.Client, plan)
 	if err != nil {
@@ -200,27 +200,10 @@ func (r *PlanReconciler) Reconcile(ctx context.Context, req ctrl.Request) (_ ctr
 		UserID:             userID,
 		Logger:             log,
 	}
-
-	err = r.syncHostConf(ctx, scope, plan)
+	patchHelper, err := patch.NewHelper(plan, r.Client)
 	if err != nil {
-		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create host conf configMap %s failed: %s", plan.Spec.HostConfName, err.Error())
-		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
+		return ctrl.Result{}, err
 	}
-	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create host conf configMap %s success", plan.Spec.HostConfName)
-
-	err = r.syncVarsConf(ctx, scope, plan)
-	if err != nil {
-		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create vars conf configMap %s failed: %s", plan.Spec.VarsConfName, err.Error())
-		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
-	}
-	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create vars conf configMap %s success", plan.Spec.VarsConfName)
-
-	err = r.syncOpsCluster(ctx, scope, plan)
-	if err != nil {
-		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create ansible cluster %s failed: %s", plan.Spec.ClusterName, err.Error())
-		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
-	}
-	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create ansible cluster %s success", plan.Spec.ClusterName)
 
 	if plan.ObjectMeta.DeletionTimestamp.IsZero() {
 		// The object is not being deleted, so if it does not have our finalizer,
@@ -264,6 +247,31 @@ func (r *PlanReconciler) Reconcile(ctx context.Context, req ctrl.Request) (_ ctr
 		return reconcile.Result{}, err
 	}
 
+	err = r.syncHostConf(ctx, &log, plan)
+	if err != nil {
+		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create host conf configMap %s failed: %s", plan.Spec.HostConfName, err.Error())
+		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
+	}
+	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create host conf configMap %s success", plan.Spec.HostConfName)
+
+	err = r.syncVarsConf(ctx, &log, plan)
+	if err != nil {
+		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create vars conf configMap %s failed: %s", plan.Spec.VarsConfName, err.Error())
+		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
+	}
+	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create vars conf configMap %s success", plan.Spec.VarsConfName)
+
+	err = r.syncOpsCluster(ctx, &log, plan)
+	if err != nil {
+		r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create ansible cluster %s failed: %s", plan.Spec.ClusterName, err.Error())
+		return reconcile.Result{RequeueAfter: RequeueAfter}, nil
+	}
+	r.EventRecorder.Eventf(plan, corev1.EventTypeNormal, PlanCreatedEvent, "Create ansible cluster %s success", plan.Spec.ClusterName)
+
+	if plan.Spec.MachineExist {
+		return reconcile.Result{}, err
+	}
+
 	defer func() {
 		if !deletion {
 			if err := patchHelper.Patch(ctx, plan); err != nil {
@@ -274,7 +282,6 @@ func (r *PlanReconciler) Reconcile(ctx context.Context, req ctrl.Request) (_ ctr
 		}
 
 	}()
-
 	// Handle non-deleted clusters
 	return r.reconcileNormal(ctx, scope, plan)
 }
@@ -440,7 +447,7 @@ func (r *PlanReconciler) SetOwnerReferences(objectMetaData *metav1.ObjectMeta, p
 	objectMetaData.OwnerReferences = []metav1.OwnerReference{*metav1.NewControllerRef(plan, ecnsv1.GroupVersion.WithKind("Plan"))}
 }
 
-func (r *PlanReconciler) syncHostConf(ctx context.Context, scope *scope.Scope, plan *ecnsv1.Plan) error {
+func (r *PlanReconciler) syncHostConf(ctx context.Context, log *logr.Logger, plan *ecnsv1.Plan) error {
 	hostConf, err := utils.CreateHostConfConfigMap(ctx, r.Client, plan)
 	if err != nil {
 		return err
@@ -452,7 +459,7 @@ func (r *PlanReconciler) syncHostConf(ctx context.Context, scope *scope.Scope, p
 	if err != nil {
 		if apierrors.IsNotFound(err) {
 			if err = r.Client.Create(ctx, hostConf); err != nil {
-				scope.Logger.Error(err, "Create host conf configMap failed")
+				log.Error(err, "Create host conf configMap failed")
 				return err
 			}
 			return nil
@@ -460,14 +467,14 @@ func (r *PlanReconciler) syncHostConf(ctx context.Context, scope *scope.Scope, p
 	}
 	err = r.Client.Update(ctx, hostConf)
 	if err != nil {
-		scope.Logger.Error(err, "Update host conf configMap failed")
+		log.Error(err, "Update host conf configMap failed")
 		return err
 	}
 
 	return nil
 }
 
-func (r *PlanReconciler) syncVarsConf(ctx context.Context, scope *scope.Scope, plan *ecnsv1.Plan) error {
+func (r *PlanReconciler) syncVarsConf(ctx context.Context, log *logr.Logger, plan *ecnsv1.Plan) error {
 	varsConf, err := utils.CreateVarsConfConfigMap(ctx, r.Client, plan)
 	if err != nil {
 		return err
@@ -480,7 +487,7 @@ func (r *PlanReconciler) syncVarsConf(ctx context.Context, scope *scope.Scope, p
 	if err != nil {
 		if apierrors.IsNotFound(err) {
 			if err = r.Client.Create(ctx, varsConf); err != nil {
-				scope.Logger.Error(err, "Create vars conf configMap failed")
+				log.Error(err, "Create vars conf configMap failed")
 				return err
 			}
 			return nil
@@ -489,31 +496,35 @@ func (r *PlanReconciler) syncVarsConf(ctx context.Context, scope *scope.Scope, p
 
 	err = r.Client.Update(ctx, varsConf)
 	if err != nil {
-		scope.Logger.Error(err, "Update vars conf configMap failed")
+		log.Error(err, "Update vars conf configMap failed")
 		return err
 	}
 
 	return nil
 }
 
-func (r *PlanReconciler) syncOpsCluster(ctx context.Context, scope *scope.Scope, plan *ecnsv1.Plan) error {
+func (r *PlanReconciler) syncOpsCluster(ctx context.Context, log *logr.Logger, plan *ecnsv1.Plan) error {
 	cluster := utils.CreateOpsCluster(plan)
-	r.SetOwnerReferences(&cluster.ObjectMeta, plan)
 	cl := &clusterv1alpha1.Cluster{}
 	err := r.Client.Get(ctx, types.NamespacedName{Namespace: cluster.Namespace, Name: cluster.Name}, cl)
 	if err != nil {
 		if apierrors.IsNotFound(err) {
+			r.SetOwnerReferences(&cluster.ObjectMeta, plan)
 			if err = r.Client.Create(ctx, cluster); err != nil {
-				scope.Logger.Error(err, "Create cluster failed")
+				log.Error(err, "Create cluster failed")
 				return err
 			}
 			return nil
 		}
 	}
 
-	err = r.Client.Update(ctx, cluster)
+	cl.Spec.HostsConfRef = cluster.Spec.HostsConfRef
+	cl.Spec.VarsConfRef = cluster.Spec.VarsConfRef
+	cl.Spec.SSHAuthRef = cluster.Spec.SSHAuthRef
+
+	err = r.Client.Update(ctx, cl)
 	if err != nil {
-		scope.Logger.Error(err, "Update cluster failed")
+		log.Error(err, "Update cluster failed")
 		return err
 	}
 
@@ -1356,42 +1367,50 @@ func (r *PlanReconciler) SetupWithManager(mgr ctrl.Manager, options controller.O
 }
 
 func (r *PlanReconciler) deletePlanResource(ctx context.Context, scope *scope.Scope, plan *ecnsv1.Plan) error {
-	// List all machineset for this plan
-	err := deleteHA(scope, plan)
+	if !plan.Spec.MachineExist {
+		// List all machineset for this plan
+		err := deleteHA(scope, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteCluster(ctx, r.Client, scope, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteMachineTemplate(ctx, r.Client, scope, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteCloudInitSecret(ctx, r.Client, scope, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteSSHKeySecert(ctx, scope, r.Client, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteAppCre(ctx, scope, r.Client, plan)
+		if err != nil {
+			return err
+		}
+
+		err = deleteSeverGroup(ctx, r.Client, scope, plan)
+		if err != nil {
+			return err
+		}
+	}
+
+	err := deleteKubeanCluster(ctx, r.Client, scope, plan)
 	if err != nil {
 		return err
 	}
 
 	err = deleteClusterOperationSets(ctx, r.Client, plan)
-	if err != nil {
-		return err
-	}
-
-	err = deleteCluster(ctx, r.Client, scope, plan)
-	if err != nil {
-		return err
-	}
-
-	err = deleteMachineTemplate(ctx, r.Client, scope, plan)
-	if err != nil {
-		return err
-	}
-
-	err = deleteCloudInitSecret(ctx, r.Client, scope, plan)
-	if err != nil {
-		return err
-	}
-
-	err = deleteSSHKeySecert(ctx, scope, r.Client, plan)
-	if err != nil {
-		return err
-	}
-	err = deleteAppCre(ctx, scope, r.Client, plan)
-	if err != nil {
-		return err
-	}
-
-	err = deleteSeverGroup(scope, plan)
 	if err != nil {
 		return err
 	}
@@ -1406,6 +1425,11 @@ func deleteClusterOperationSets(ctx context.Context, cli client.Client, plan *ec
 	if err != nil {
 		return err
 	}
+
+	if len(sets.Items) <= 0 {
+		return nil
+	}
+
 	for index, set := range sets.Items {
 		if set.Annotations[ecnsv1.KubeanAnnotation] != plan.Spec.ClusterName {
 			continue
@@ -1651,3 +1675,24 @@ func deleteSSHKeySecert(ctx context.Context, scope *scope.Scope, client client.C
 
 	return nil
 }
+
+func deleteKubeanCluster(ctx context.Context, client client.Client, scope *scope.Scope, plan *ecnsv1.Plan) error {
+	cl := &clusterv1alpha1.Cluster{}
+	err := client.Get(ctx, types.NamespacedName{Namespace: plan.Namespace, Name: plan.Spec.ClusterName}, cl)
+	if err != nil {
+		if apierrors.IsNotFound(err) {
+			scope.Logger.Error(err, "kubean cluster has already been deleted")
+			return nil
+		}
+		return err
+	}
+
+	err = client.Delete(ctx, cl)
+	if err != nil {
+		scope.Logger.Error(err, "Delete cluster failed")
+		return err
+	}
+
+	return nil
+}
+
